+++ src/mink/tasks/__init__.py
@@ -1,5 +1,6 @@
 """Kinematic tasks."""
 
+from .axis_align_task import AxisAlignTask as AxisAlignTask
 from .com_task import ComTask as ComTask
 from .damping_task import DampingTask as DampingTask
 from .dof_freezing_task import DofFreezingTask as DofFreezingTask